Made a successful network call to CreateBudgetLineItem Endpoint

+++ app/src/main/java/com/decagonhq/decapay/feature/createbudgetlineitems/data/network/model/createbudgetlineitemmodel/CreateBudgetLineItemRequestBody.kt
@@ -1,6 +1,6 @@
 package com.decagonhq.decapay.feature.createbudgetlineitems.data.network.model.createbudgetlineitemmodel
 
 data class CreateBudgetLineItemRequestBody(
-    val amount: Int?,
+    val amount: Double?,
     val budgetCategoryId: Int?
 )
